android: Only publish/receive video if setting SDP doesn't fail
A bad SDP will be rejected by the WebRTC implementation, so no further
publish/subscribe operations should be done on failure.

This expands on the idea of
https://github.com/OpenVidu/openvidu-tutorials/pull/174
and also simplifies implementations of CustomSdpObserver: there is no
need to inherit its methods just for logging, as CustomSdpObserver
already logs everything on its default implementation.